Finally use adaptive scaling that works decently
There's a gap between the bottom of .content and the footer anchored to
the bottom that grows when the page is narrow, but it doesn't look bad.
It also responds reasonably to the page being very wide and short,
although the footer covers the other content when you get too far.

I'm gonna go back to having decent aesthetics next.